api/garmin/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
garminconnect>=0.3.2
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
pydantic>=2.0.0
